requirements.txt
flask==3.0.0
openai>=1.6.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
PyMuPDF==1.23.8
faiss-cpu==1.7.4
python-dotenv==1.0.0
numpy==1.24.3
tiktoken>=0.5.2
gunicorn==21.2.0

